Split Error module out of Cardano.Api.Typed
